fixed service and controller
