+++ rides-service/src/main/java/com/java/akdev/ridesservice/controller/RideController.java
@@ -3,12 +3,12 @@ package com.java.akdev.ridesservice.controller;
 import com.java.akdev.ridesservice.dto.RideCreateDto;
 import com.java.akdev.ridesservice.dto.RideReadDto;
 import com.java.akdev.ridesservice.dto.RideUpdateDto;
+import com.java.akdev.ridesservice.enumeration.Order;
 import com.java.akdev.ridesservice.enumeration.SortField;
 import com.java.akdev.ridesservice.service.RideService;
 import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -23,7 +23,7 @@ public class RideController {
     public ResponseEntity<Page<RideReadDto>> findAll(@RequestParam @Min(value = 1) Integer page,
                                                      @RequestParam @Min(value = 1) Integer size,
                                                      @RequestParam SortField sortField,
-                                                     @RequestParam Sort.Direction order) {
+                                                     @RequestParam Order order) {
         return ResponseEntity.status(200)
                 .body(rideService.findAll(page, size, sortField, order));
     }
