+++ rides-service/src/main/java/com/java/akdev/ridesservice/enumeration/Order.java
@@ -1,0 +1,5 @@
+package com.java.akdev.ridesservice.enumeration;
+
+public enum Order {
+    ASC, DESC
+}
